Add .yardopts + namespace errors + move regexes
