+++ .yardopts
@@ -1,0 +1,4 @@
+--no-private
+-
+CHANGELOG.md
+LICENSE
